Modified parameters name, type of function and add gitignore file.

+++ Assignment1/src/main/java/comp3011/assignment/components/schemas/GlobalStatsResponse.java
@@ -1,4 +1,4 @@
 package comp3011.assignment.components.schemas;
 
-public record GlobalStatsResponse(int inputTokens, int outputTokens) 
+public record GlobalStatsResponse(long inputTokens, long outputTokens) 
 {}
